[ALPS08678993] mml:build error fix
Fix 32bit sys build error.

MTK-Commit-Id: bc77bfb59c98f77e05a55631e3a1c751b2629252

CR-Id: ALPS08678993
Feature: mml

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-driver.c
@@ -808,7 +808,7 @@ s32 mml_subcomp_init(struct platform_device *comp_pdev,
 		ret = snprintf(name, sizeof(name), "%s-clock-names", comp->name);
 		if (ret >= sizeof(name)) {
 			dev_err(dev, "len:%d over name size:%lu",
-				ret, sizeof(name));
+				ret, (unsigned long)sizeof(name));
 			name[sizeof(name) - 1] = '\0';
 		}
 		name_ptr = name;
